+++ azext_iot/iothub/providers/mqtt.py
@@ -4,6 +4,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
+import os
 import pprint
 from time import sleep
 from typing import Any, Dict, Optional
@@ -61,11 +62,32 @@ class MQTTProvider(object):
         self.init_reported_properties = init_reported_properties
 
     def send_d2c_message(
-        self, message_text: str, properties: Optional[Dict[str, Any]] = None
+        self,
+        message_content: str,
+        message_file_path: Optional[str] = None,
+        properties: Optional[Dict[str, Any]] = None
     ):
         from azure.iot.device import Message
 
-        message = Message(message_text)
+        if message_file_path:
+            if not os.path.exists(message_file_path):
+                raise FileNotFoundError("File path does not exist: {}".format(message_file_path))
+
+            binary_content = False
+            if properties:
+                binary_content = properties.get("$.ct", None) == 'application/octet-stream'
+
+            # send bytes as message when content type is defined as binary
+            if binary_content:
+                with open(message_file_path, "rb") as f:
+                    data = f.read()
+            else:
+                with open(message_file_path, "r", encoding="utf-8") as f:
+                    data = f.read()
+        else:
+            data = message_content
+
+        message = Message(data)
         message.custom_properties = properties
         self.device_client.send_message(message)
 
@@ -159,7 +181,7 @@ class MQTTProvider(object):
                 self.device_client.patch_twin_reported_properties(self.init_reported_properties)
 
             for _ in tqdm(range(0, msg_count), desc='Device simulation in progress', ascii=' #'):
-                self.send_d2c_message(message_text=data.generate(True), properties=properties)
+                self.send_d2c_message(message_content=data.generate(True), properties=properties)
                 sleep(publish_delay)
 
         except Exception as x: