+++ HISTORY.rst
@@ -7,6 +7,26 @@ Release History
 unreleased
 +++++++++++++++
 
+**IoT Hub updates**
+
+* Fix `az iot hub monitor-events` hangs when message payload failed to find representation in unicode.
+
+**IoT device updates**
+
+* `az iot device send-d2c-message` now support inserting file path with `--data-file-path` argument.
+* `az iot device c2d-message send` now support inserting file path with `--data-file-path` argument.
+* `az iot device c2d-message receive` now shows non-decodable message payload with rest of the properties.
+
+
+0.21.1
++++++++++++++++
+
+**IoT Hub updates**
+
+* Improved help for `az iot edge deployment create` to better describe layered deployments.
+
+* The command `az iot edge export-modules` is now GA.
+
 
 0.21.0
 +++++++++++++++